Working on separation of concerns and moving game logic to gamecontroller.
I have maintained state for storing of characters as this is referenced throughout the document for conditional formatting.

I've maintained state for tracking of timers etc as these require useEffects to work which require React.

Need to review if any other game logic can be moved to gameController, then continue with project.